lora_trainer/requirements.txt
accelerate==0.34.2
datasets==3.0.1
diffusers==0.30.3
numpy==2.1.1
Pillow==10.4.0
PyYAML==6.0.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.0
